Optimized brands seo on frontpage into component, same for car part categories seo and finished and cleaned up code with comments.
