+++ resources/views/components/brands/categories.blade.php
