+++ routes/web.php
@@ -75,8 +75,6 @@ Route::group([
     Route::get('/categories/{name}/{id}/subcategories', [LandingPageController::class, 'showSubCategories'])->name('categories.show');
     Route::get('/subcategories/{name}/{id}/brands', [LandingPageController::class, 'showBrandsForSubcategories'])->name('subcategories.brands');
     Route::get('/subcategories/{subCategoryName}/{subCategoryId}/brands/{brandName}/{brandId}/models', [LandingPageController::class, 'showModelsForSubCategoryAndBrand'])->name('subcategories.brands.models');
-    Route::get('/subcategories/{subCategoryName}/{subCategoryId}/brands/{brandName}/{brandId}/models/{modelName}/{modelId}/search', [LandingPageController::class, 'searchCarParts'])->name('subcategories.brands.models.search');
-
     // Payment routes
     Route::post('products/{carPart}/payments/pay', [PaymentController::class, 'pay'])
         ->name('pay');
